requirements.txt
django==3.0.7
psycopg2>=2.7,<3.0
requests==2.24.0
Pillow==7.1.2